Move test and assert static imports inserted
